{"schema":"decodex/commit/1","summary":"Extract frozen export model owner","authority":"manual"}
