+++ docs/reference/workspace-layout.md
@@ -120,6 +120,9 @@ Key paths:
   `overlay/session_state/scroll_capture.rs`
 - `packages/rsnap-overlay/src/frozen_edit.rs`: Rust-owned frozen overlay edit session state machine;
   edit geometry, text hit bounds, and movement clamping live under `frozen_edit/geometry.rs`
+- `packages/rsnap-overlay/src/frozen_export.rs`: Rust-owned frozen-overlay export compositor;
+  export element schema lives under `frozen_export/model.rs`, and stroke rasterization lives under
+  `frozen_export/stroke_raster.rs`
 - `packages/rsnap-overlay/src/overlay/frozen_brush_runtime.rs`: Frozen pen interaction routing
   for `OverlaySession`; the brush sampling, response, smoothing, preview, and rendered point model
   lives under `overlay/frozen_brush_model.rs`
